Safe ast max_workspace_size to int

+++ mmdet2trt/mmdet2trt.py
@@ -97,7 +97,7 @@ def main():
     )
     parser.add_argument(
         "--max-workspace-gb",
-        type=int,
+        type=float,
         default=0.5,
         help="The maximum `device` (GPU) temporary memory in GB (gigabytes) which TensorRT can use at execution time.",
     )
@@ -162,7 +162,7 @@ def main():
         args.checkpoint,
         device=args.device,
         fp16_mode=args.fp16,
-        max_workspace_size=args.max_workspace_gb * 1e9,
+        max_workspace_size=int(args.max_workspace_gb * 1e9),
         opt_shape_param=opt_shape_param,
         trt_log_level=args.trt_log_level,
         return_warp_model=args.return_warp_model,
